+++ idg-lib/src/CPU/Reference/kernels/KernelSplitter.cpp
@@ -54,8 +54,8 @@ void kernel_splitter(
                     // Set grid value to subgrid
                     for (int pol = 0; pol < NR_POLARIZATIONS; pol++) {
                         int pol_src  = index_pol[pol];
-                        int src_idx = index_grid(grid_size, pol_src, y_src, x_src);
-                        int dst_idx = index_subgrid(NR_POLARIZATIONS, subgrid_size, s, pol, y_dst, x_dst);
+                        long src_idx = index_grid(grid_size, pol_src, y_src, x_src);
+                        long dst_idx = index_subgrid(NR_POLARIZATIONS, subgrid_size, s, pol, y_dst, x_dst);
                         idg::float2 value = grid[src_idx];
                         value = negative_w ? conj(value) : value;
                         subgrid[dst_idx] = phasor * value;
